Renamed the examples submodule to datasets and fixed some smaller bugs.
